pyproject.toml
[project]
name = "personal-finance-tracker"
version = "0.1.0"
description = "Personal portfolio tracking system with SQLite, Python, and spreadsheet UI"
readme = "README.md"
requires-python = ">=3.14"
authors = [
    {name = "Keith"}
]

dependencies = [
    "tradingview-scraper>=0.1.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-api-python-client>=2.0.0",
    "python-dotenv>=1.0.0",
    "ccxt>=4.0.0",  # Crypto exchange integration
    "web3"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py
